// products/affiliate-sites/fragrance-dupes/lib/data/pm-offers.generated.ts
export const PM_MERCHANT = {
  name: "Perfumania.com",
  advertiserId: "17335854",
  domain: "perfumania.com",
} as const;